Refine Users, Events and OfflineListener services
