+++ client/app/services/ng.js
@@ -1,10 +1,12 @@
 export let $http = null; // eslint-disable-line import/no-mutable-exports
+export let $sanitize = null; // eslint-disable-line import/no-mutable-exports
 export let $uibModal = null; // eslint-disable-line import/no-mutable-exports
 export let toastr = null; // eslint-disable-line import/no-mutable-exports
 
 export default function init(ngModule) {
   ngModule.run(($injector) => {
     $http = $injector.get('$http');
+    $sanitize = $injector.get('$sanitize');
     $uibModal = $injector.get('$uibModal');
     toastr = $injector.get('toastr');
   });
